refactor: 값으로써의 역할 보다, 계층 간 이동을 담당하는 dto의 역할에 가까운 클래스의 패키지를 이동
